Add inventory adjustment
