+++ client/packages/system/src/Stock/Components/InventoryAdjustment/index.ts
@@ -1,1 +1,1 @@
-export * from './InventoryAdjustmentForm';
+export * from './InventoryAdjustmentModal';
