+++ client/packages/system/src/Stock/Components/index.ts
@@ -1,1 +1,2 @@
 export * from './Repack';
+export * from './InventoryAdjustment';
